Removed unnecessary print statements
The print statement on line 101 has been left in case the user wishes to see the nbt tag of what they are editing. This is especially useful if they wish to see the nbt tag they are editing. 

A possible addition would be to use difflib to print (or even log to a file) the difference between any entities or tile entities that have been changed.